+++ src/main/java/com/zc/service/ZcApiService.java
@@ -5,7 +5,7 @@ import com.zc.bean.ZcApi;
 import java.util.List;
 
 public interface ZcApiService {
-    int insertZcApi(ZcApi object);
+    Long insertZcApi(ZcApi object);
 
     int updateZcApi(ZcApi object);
