+++ src/main/java/com/zc/service/ZcCashOutRecordService.java
@@ -5,7 +5,7 @@ import com.zc.bean.ZcCashOutRecord;
 import java.util.List;
 
 public interface ZcCashOutRecordService {
-    int insertZcCashOutRecord(ZcCashOutRecord object);
+    Long insertZcCashOutRecord(ZcCashOutRecord object);
 
     int updateZcCashOutRecord(ZcCashOutRecord object);
